+++ src/api.js
@@ -97,6 +97,206 @@
             loggedInToken = token;
         },
         /**
+         * Adds a collaborator to the given repository.
+         *
+         * @param {String} username - the username to add a collaborator to
+         * @param {String} name - the name of the repository to add a collaborator to
+         * @param {String} collaborator - the username of the collaborator to add
+         * @returns {Promise}
+         */
+        addCollaborator: function (username, name, collaborator) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!collaborator || typeof collaborator !== 'string') {
+                    return reject(new Error('Collaborator username must be provided!'));
+                }
+
+                // Make sure the username/collaborator is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+                collaborator = collaborator.toLowerCase();
+
+                this.makePostRequest(`repositories/${username}/${name}/collaborators/`, {user: collaborator}).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the details for a given build of a repository.
+         *
+         * @param {String} username - the username to get the comments for
+         * @param {String} name - the name of the repository to get the comments for
+         * @param {String} code - the code of the build to get the details for
+         * @returns {Promise}
+         */
+        buildDetails: function (username, name, code) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!code || typeof code !== 'string') {
+                    return reject(new Error('Repository code must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/buildhistory/${code}`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Get the history of all the builds done for a given repository.
+         *
+         * @param {String} username - the username to get the build history for
+         * @param {String} name - the name of the repository to get the build history for
+         * @param {{page: Number, perPage: Number}} [options] - the options for this call
+         * @returns {Promise}
+         */
+        buildHistory: function (username, name, options) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!options) {
+                    options = {page: 1, perPage: 100};
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/buildhistory?page_size=${options.perPage || 100}&page=${options.page || 1}`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the build links for a given repository.
+         *
+         * @param {String} username - the username to get the comments for
+         * @param {String} name - the name of the repository to get the comments for
+         * @returns {Promise}
+         */
+        buildLinks: function (username, name) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/links`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the build settings for a given repository.
+         *
+         * @param {String} username - the username to get the comments for
+         * @param {String} name - the name of the repository to get the comments for
+         * @returns {Promise}
+         */
+        buildSettings: function (username, name) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/autobuild`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the build trigger for a given repository.
+         *
+         * @param {String} username - the username to get the comments for
+         * @param {String} name - the name of the repository to get the comments for
+         * @returns {Promise}
+         */
+        buildTrigger: function (username, name) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/buildtrigger`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the build trigger history for a given repository.
+         *
+         * @param {String} username - the username to get the comments for
+         * @param {String} name - the name of the repository to get the comments for
+         * @returns {Promise}
+         */
+        buildTriggerHistory: function (username, name) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/buildtrigger/history`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Gets the collaborators for a repository.
+         *
+         * @param {String} username - the username to get the collaborators for
+         * @param {String} name - the name of the repository to get the collaborators for
+         * @returns {Promise}
+         */
+        collaborators: function (username, name) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeGetRequest(`repositories/${username}/${name}/collaborators`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
          * Gets the comments for a repository.
          *
          * @param {String} username - the username to get the comments for
@@ -134,6 +334,129 @@
                 username = username.toLowerCase();
 
                 this.makeGetRequest(`repositories/${username}/${name}/comments?page_size=${options.perPage || 100}&page=${options.page || 1}`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Creates a build link for a given repository to the given repository.
+         *
+         * @param {String} username - the username to create a build link for
+         * @param {String} name - the name of the repository to create a build link for
+         * @param {String} to_repo - the repo to link this automated build to
+         * @returns {Promise}
+         */
+        createBuildLink: function (username, name, to_repo) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!to_repo || typeof to_repo !== 'string') {
+                    return reject(new Error('Repo to link to must be provided!'));
+                }
+
+                // Check to see if the user provided a username for the to_repo
+                if (to_repo.indexOf('/') === -1) {
+                    to_repo = 'library/' + to_repo;
+                }
+
+                // If to-repo has '_/' then we're trying to get an official repository
+                if (to_repo.substr(0, 2) === '_/') {
+                    to_repo = 'library/' + to_repo.substr(2);
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makePostRequest(`repositories/${username}/${name}/links`, {to_repo}).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Creates a build tag for a given repository.
+         *
+         * @param {String} username - the username to create a build tag for
+         * @param {String} name - the name of the repository to create a build tag for
+         * @param {Object} details - the details of the build tag
+         * @returns {Promise}
+         */
+        createBuildTag: function (username, name, details) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!details || typeof details !== 'object') {
+                    return reject(new Error('Tag details must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                // Build our object to post
+                let obj = {
+                    isNew: true,
+                    namespace: username,
+                    repoName: name,
+                    name: details.name || 'latest',
+                    dockerfile_location: details.dockerfile_location || '/',
+                    source_type: details.source_type || 'Branch',
+                    source_name: details.source_name || 'master'
+                };
+
+                this.makePostRequest(`repositories/${username}/${name}/autobuild/tags`, obj).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Creates an automated build.
+         *
+         * @param {String} username - the username of the automated build to create
+         * @param {String} name - the name of the automated build to create
+         * @param {Object} details - the details of the new automated build
+         * @returns {Promise}
+         */
+        createAutomatedBuild: function (username, name, details) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!details || typeof details !== 'object') {
+                    return reject(new Error('Details must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                let obj = {
+                    name,
+                    namespace: username,
+                    active: true,
+                    dockerhub_repo_name: `${username}/${name}`,
+                    provider: details.provider,
+                    vcs_repo_name: details.vcs_repo_name,
+                    is_private: false
+                };
+
+                if (typeof details.is_private === 'boolean') {
+                    obj.is_private = details.is_private;
+                }
+
+                if (details.description) {
+                    obj.description = details.description;
+                }
+
+                return this.makePostRequest(`repositories/${username}/${name}/autobuild/`, obj).then(resolve).catch(reject);
             }.bind(this));
         },
         /**
@@ -243,6 +566,99 @@
             }.bind(this));
         },
         /**
+         * Deletes a build link for a given repository.
+         *
+         * @param {String} username - the username to delete a build link for
+         * @param {String} name - the name of the repository to delete a build link for
+         * @param {Number} id - the id of the build link to delete
+         * @returns {Promise}
+         */
+        deleteBuildLink: function (username, name, id) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!id) {
+                    return reject(new Error('Build link id must be provided!'));
+                }
+
+                if (typeof id !== 'number' || id < 0) {
+                    return reject(new Error('Build link id must be a number greater than 0!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeDeleteRequest(`repositories/${username}/${name}/links/${id}`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Deletes a build tag for a given repository.
+         *
+         * @param {String} username - the username to delete a build tag for
+         * @param {String} name - the name of the repository to delete a build tag for
+         * @param {Number} id - the id of the build tag to delete
+         * @returns {Promise}
+         */
+        deleteBuildTag: function (username, name, id) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!id) {
+                    return reject(new Error('Build link id must be provided!'));
+                }
+
+                if (typeof id !== 'number' || id < 0) {
+                    return reject(new Error('Build link id must be a number greater than 0!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                this.makeDeleteRequest(`repositories/${username}/${name}/autobuild/tags/${id}`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * Deletes a build tag for a given repository.
+         *
+         * @param {String} username - the username to delete a collaborator for
+         * @param {String} name - the name of the repository to delete a collaborator for
+         * @param {String} collaborator - the username of the collaborator to delete
+         * @returns {Promise}
+         */
+        deleteCollaborator: function (username, name, collaborator) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!collaborator || typeof collaborator !== 'string') {
+                    return reject(new Error('Collaborator\'s username must be provided!'));
+                }
+
+                // Make sure the username and collaborator is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+                collaborator = collaborator.toLowerCase();
+
+                this.makeDeleteRequest(`repositories/${username}/${name}/collaborators/${collaborator}`).then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
          * Deletes a repository.
          *
          * @param {String} username - the username of the repository to delete
@@ -322,6 +738,19 @@
             }.bind(this));
         },
         /**
+         * This gets the registry settings for the current logged in user containing information about the number of private repositories used/available.
+         *
+         * @returns {Promise}
+         */
+        registrySettings: function () {
+            return new Promise(function (resolve, reject) {
+                this.loggedInUser().then(function (user) {
+                    return this.makeGetRequest(`users/${user.username}/registry-settings`);
+                }.bind(this)).then(resolve).catch(reject);
+
+            }.bind(this));
+        },
+        /**
          * Gets the details about a repository.
          *
          * @param {String} [username] - the username of the repository to get information about. If left out or '_' is provided then it will query the official Docker repository with the given name
@@ -384,6 +813,49 @@
                 username = username.toLowerCase();
 
                 this.makeGetRequest(`users/${username}/repositories/starred?page_size=${options.perPage || 100}&page=${options.page || 1}`, 'results').then(resolve).catch(reject);
+            }.bind(this));
+        },
+        /**
+         * This sets the details of a build tag for a given users repository.
+         *
+         * @param {String} username - the username
+         * @param {String} name - the name of the repository
+         * @param {Number} id - the id of the build tag to save
+         * @param {Object} details - an object with a the new details of the build tag
+         * @returns {Promise}
+         */
+        setBuildTag: function (username, name, id, details) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!id) {
+                    return reject(new Error('id must be provided!'));
+                }
+
+                if (typeof id !== 'number' || id <= 0) {
+                    return reject(new Error('Provided id must be a number greater than 0!'));
+                }
+
+                if (!details || typeof details !== 'object') {
+                    return reject(new Error('Tag details must be provided!'));
+                }
+
+                // Build our object to post
+                let obj = {
+                    id,
+                    name: details.name || 'latest',
+                    dockerfile_location: details.dockerfile_location || '/',
+                    source_type: details.source_type || 'Branch',
+                    source_name: details.source_name || 'master'
+                };
+
+                return this.makePutRequest(`repositories/${username}/${name}/autobuild/tags/${id}`, obj).then(resolve).catch(reject);
             }.bind(this));
         },
         /**
@@ -475,6 +947,41 @@
             username = username.toLowerCase();
 
             return this.makeGetRequest(`repositories/${username}/${name}/tags?page_size=${options.perPage || 100}&page=${options.page || 1}`, 'results');
+        },
+        /**
+         * Creates a build tag for a given repository.
+         *
+         * @param {String} username - the username to create a build tag for
+         * @param {String} name - the name of the repository to create a build tag for
+         * @param {Object} details - the details of the build tag
+         * @returns {Promise}
+         */
+        triggerBuild: function (username, name, details) {
+            return new Promise(function (resolve, reject) {
+                if (!username || typeof username !== 'string') {
+                    return reject(new Error('Username must be provided!'));
+                }
+
+                if (!name || typeof name !== 'string') {
+                    return reject(new Error('Repository name must be provided!'));
+                }
+
+                if (!details || typeof details !== 'object') {
+                    return reject(new Error('Build details must be provided!'));
+                }
+
+                // Make sure the username is all lowercase as per Docker Hub requirements
+                username = username.toLowerCase();
+
+                // Build our object to post
+                let obj = {
+                    dockerfile_location: details.dockerfile_location || '/',
+                    source_type: details.source_type || 'Branch',
+                    source_name: details.source_name || 'master'
+                };
+
+                this.makePostRequest(`repositories/${username}/${name}/autobuild/trigger-build`, obj).then(resolve).catch(reject);
+            }.bind(this));
         },
         /**
          * This unstars a repository for the logged in user.
@@ -663,6 +1170,48 @@
                 }
 
                 request(this.makeRequestParams('post', path, data), function (err, res, body) {
+                    if (err) {
+                        return reject(err);
+                    }
+
+                    // Some api calls don't return any data
+                    if (!body) {
+                        return resolve();
+                    }
+
+                    // If the body has a detail property, it's only because there's an error I've found
+                    if (body.detail) {
+                        return reject(new Error(body.detail));
+                    }
+
+                    // If the body has a error property, then it's errored out
+                    if (body.error) {
+                        return reject(new Error(body.error));
+                    }
+
+                    if (extract && body.hasOwnProperty(extract)) {
+                        return resolve(body[extract]);
+                    }
+
+                    return resolve(body);
+                });
+            }.bind(this));
+        },
+        /**
+         * Makes a raw put request to the Docker Hub API.
+         *
+         * @param {String} path - the path to fetch
+         * @param {Object} data - the data to send
+         * @param {String} [extract] - the name of the property in the resulting JSON to extract. If left blank it will return the entire JSON
+         * @returns {Promise}
+         */
+        makePutRequest(path, data, extract) {
+            return new Promise(function (resolve, reject) {
+                if (!data || typeof data !== 'object') {
+                    return reject(new Error('Data must be passed to all PUT requests in the form of an object!'));
+                }
+
+                request(this.makeRequestParams('put', path, data), function (err, res, body) {
                     if (err) {
                         return reject(err);
                     }